Move land tool functions from TopToolBar.cpp into Land.cpp
